% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-cedar_rapids.r
\docType{data}
\name{cedar_rapids}
\alias{cedar_rapids}
\title{Cedar Rapids data set (1903-2017)}
\format{
A data frame with n rows and 2 variables

\describe{
\item{year}{year}
\item{riverflow}{cubic feet per second}
}
}
\source{
\url{https://waterdata.usgs.gov}
}
\usage{
cedar_rapids
}
\description{
Annual riverflow discharge (cubic feet per second) from Cedar Rapids, IA  (USGS station #05464500)
}
\section{Data download}{

}

\seealso{
Other data: 
\code{\link{bottom_temp_survey}},
\code{\link{surface_temp_oisst}}
}
\concept{data}
\keyword{datasets}
