% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data_validation.R
\name{check_data_validation}
\alias{check_data_validation}
\title{Checks data set for misspecification}
\usage{
check_data_validation(dataSet)
}
\description{
Check column names, missing data, consecutive equally spaced time units
}
