% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_beta_given_rho.R
\name{est_beta_given_rho}
\alias{est_beta_given_rho}
\title{Estimate beta conditional on known AR1 parameter}
\usage{
est_beta_given_rho(xt, yt, rho)
}
\arguments{
\item{xt}{Numeric vector. Time series of time points (1, ..., n)}

\item{yt}{Numeric vector. Time series of observations (length = n)}

\item{rho}{Numeric scalar. AR1 parameter}
}
\description{
Uses Prais and Winsten transformation which utilizes first observation
}
