% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_beta_given_rho2.R
\name{est_beta_given_rho2}
\alias{est_beta_given_rho2}
\title{Estimate beta conditional on known AR1 parameter}
\usage{
est_beta_given_rho2(xt, yt, rho1, rho2 = 0)
}
\description{
Uses Prais and Winsten transformation which utilizes first observation
}
