% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_bottom_temp.R
\name{example_bottom_temp}
\alias{example_bottom_temp}
\title{Run test on bottom temperature example data set}
\usage{
example_bottom_temp(epu = "GB", nBootSims = 999)
}
\arguments{
\item{epu}{character string. Name of the EPU to test ("GB", "GOM")}

\item{nBootSims}{Numeric scalar. Number of bootstrap samples to perform}
}
\description{
Performs the short term trend analysis on bundled data set
}
