% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_cedar_rapids.R
\name{example_cedar_rapids}
\alias{example_cedar_rapids}
\title{Run test on example data set}
\usage{
example_cedar_rapids(dataSet = arfit::cedar_rapids, nBootSims = 999)
}
\arguments{
\item{dataSet}{Data frame. Two columns (year, riverflow)}

\item{nBootSims}{Numeric scalar. Number of bootstrap samples to perform}
}
\description{
Use data from USGS 05464500 Cedar River at Cedar Rapids, IA
}
