% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_opt.R
\name{example_opt}
\alias{example_opt}
\title{Run test on example data set}
\usage{
example_opt(dataSet, ylabel, nBootSims = 999)
}
\arguments{
\item{dataSet}{Data frame. Two columns (x, y)}

\item{nBootSims}{Numeric scalar. Number of bootstrap samples to perform}
}
\description{
Supply data
}
