% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_surface_temp.R
\name{example_surface_temp}
\alias{example_surface_temp}
\title{Run test on example OISST data set}
\usage{
example_surface_temp(epu = "GB", nBootSims = 999)
}
\arguments{
\item{epu}{character string. Name of the EPU to test ("GB", "GOM")}

\item{nBootSims}{Numeric scalar. Number of bootstrap samples to perform}
}
\description{
Performs the short term trend analysis on bundled data set
}
