% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ar1_opt.R
\name{fit_ar1_opt}
\alias{fit_ar1_opt}
\title{Fit linear model with trend using optimization routine}
\usage{
fit_ar1_opt(data, rho, hypothesis)
}
\description{
Fit linear model with trend using optimization routine
}
