% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_real_data.R
\name{fit_real_data}
\alias{fit_real_data}
\title{Run test on real data set}
\usage{
fit_real_data(dataSet, nBootSims = 499, printFig = F)
}
\arguments{
\item{dataSet}{Data frame.
Two of the fields must be named x (time, equally spaced), and y (response variable)}

\item{nBootSims}{Numeric scalar. Number of bootstrap samples to perform}

\item{printFig}{Boolean. Print data an fit in figure window (Default = F)}
}
\value{
A list containing:
\item{null}{Fitted model under the null hypothesis (no trend)}
\item{alt}{Fitted model under the alternative hypothesis (with trend)}
\item{pValue}{p-value from the bootstrap test}
\item{pValChi2}{p-value from the chi-squared approximation}
\item{data}{The input data set}
}
\description{
Performs a parametric bootstrap to test the significance of a linear trend in time under
the assumption of autoregressive order 1 error.
}
\section{}{
 ecodata

This function is used in ecodata::geom_lm()
}

