% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_to_data.R
\name{fit_to_data}
\alias{fit_to_data}
\title{Run test on user data set}
\usage{
fit_to_data(dataSet = NULL, nBootSims = 999)
}
\arguments{
\item{dataSet}{Data frame. Two columns (x,y)}

\item{nBootSims}{Numeric scalar. Number of bootstrap samples to perform}
}
\description{
Fits a linear model to data set assuming AR1 error
}
