% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isStationary_ar2.R
\name{isStationary_ar2}
\alias{isStationary_ar2}
\title{Determines stationarity of AR2}
\usage{
isStationary_ar2(rhos)
}
\arguments{
\item{rhos}{Numeric vector. AR2 parameters}
}
\value{
Boolean. Process is stationary or not
}
\description{
Given parameters of AR2 determines if process is stationary
}
