% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_ar1.R
\name{likelihood_ar1}
\alias{likelihood_ar1}
\title{Evaluate the likelihood of linear model with errors from an AR order 1 process}
\usage{
likelihood_ar1(rho, dataf, hypothesis = "null")
}
\arguments{
\item{rho}{Numeric scalar. Autoregressive parameter}

\item{dataf}{Data frame. (nx2). Defining xt and yt values}

\item{hypothesis}{Character string. "null" or "alt" determines whether to evaluate
under the null or alternative hypothesis. "null" fixes linear component, beta = 0.}
}
\value{

}
\description{
Evaluate the likelihood of linear model with errors from an AR order 1 process
}
