% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_single_opt_ar2.R
\name{sim_single_opt_ar2}
\alias{sim_single_opt_ar2}
\title{Simulates n data sets and performs bootstrapping}
\usage{
sim_single_opt_ar2(
  beta = 0,
  rho1 = 0,
  rho2 = 0,
  sigma = 0.25,
  nT = 10,
  nSims = 200,
  nBootSims = 500
)
}
\arguments{
\item{beta}{Numeric.  Values for beta_1 (slope/trend parameter)}

\item{sigma}{Numeric.  Values of standard deviation of noise}

\item{nT}{Numeric.  Values for the length of time series to simulate}

\item{nSims}{Numeric.  Number of time series to simulate}

\item{nBootSims}{Numeric.  Number of bootstrap data sets}

\item{rho}{Numeric  Values for autoregressive parameter}
}
\description{
Performs n tests
}
\details{
For convenience the intercept, beta_0 is set to zero
}
\examples{
\dontrun{
sim_single_opt(beta = 0.5,rho =0.5, sigma = 0.25, nT = 10,
nSims = 200, nBootSims = 500)
}

}
