% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_ar1.R
\name{simulate_ar1}
\alias{simulate_ar1}
\title{Simulate data from an linear process with AR1 error}
\usage{
simulate_ar1(alpha, beta = 0, sigma, rho, n)
}
\arguments{
\item{alpha}{Numeric scalar. Intercept}

\item{beta}{Numeric scalar. Slope}

\item{sigma}{Numeric scalar. Standard deviation of error term}

\item{rho}{Numeric scalar. Auto regressive parameter}

\item{n}{Numeric scalar. Length of time series}
}
\description{
model=   y_t = a + b\emph{t + z_t
where z_t = rho} z_t-1 + e_t   (e_t ~ N(0,sigma^2))
Uses unconditional mean and variance to simulate first data point
}
