% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_study_opt_ar2.R
\name{simulation_study_opt_ar2}
\alias{simulation_study_opt_ar2}
\title{Simulation study using optimization routine}
\usage{
simulation_study_opt_ar2(
  outDir = here::here("out_ar2.txt"),
  betaVec = c(0),
  rho1Vec = c(0),
  rho2Vec = c(0),
  sigmaVec = c(0.25),
  nTVec = c(10),
  nSims = 200,
  nBootSims = 500,
  rhoSearch = NULL
)
}
\arguments{
\item{outDir}{Character string. Path to output file}

\item{betaVec}{Numeric vector. Values for beta_1 (slope/trend parameter)}

\item{rho1Vec}{Numeric vector. Values for autoregressive parameter, rho_!}

\item{rho2Vec}{Numeric vector. Values for autoregressive parameter, rho_2}

\item{sigmaVec}{Numeric vector. Values of standard deviation of noise}

\item{nSims}{Numeric scalar. Number of time series to simulate}

\item{nBootSims}{Numeric scalar. Number of bootstrap data sets}

\item{nVec}{Numeric vector. Values for the length of time series to simulate}
}
\description{
Performs a simulation study to assess the performance of the test.
}
\details{
For convenience the intercept, beta_0 is set to zero
}
\examples{
\dontrun{
simulation_study_opt(here::here("output.txt"),betaVec = 0.5,rhovec =0.5, sigmaVec = 0.25,
nTVec = 10, nSims = 200, nBootSims = 500)
}

}
