% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-surface_temp_oisst.r
\docType{data}
\name{surface_temp_oisst}
\alias{surface_temp_oisst}
\title{OISST temperature data}
\format{
A data frame

\describe{
\item{Time}{year}
\item{Var}{name of the variable}
\item{Value}{value of the variable}
\item{EPU}{Ecological Production Unit in which Var was measured/calculated}
\item{Units}{units measures}
}
}
\source{
\url{https://noaa-edab.github.io/ecodata/index.html}
}
\usage{
surface_temp_oisst
}
\description{
Most recent 10 years of data from National Oceanographic and Atmospheric Administration optimum interpolation sea surface temperature high resolution data set (NOAA OISST V2)
}
\section{Data}{


The Data were extracted from ecodata v6.0.1 from the indicator \emph{seasonal_oisst_anom}
}

\seealso{
Other data: 
\code{\link{bottom_temp_survey}},
\code{\link{cedar_rapids}}
}
\concept{data}
\keyword{datasets}
