% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arfit.R
\docType{package}
\name{arfit-package}
\alias{arfit}
\alias{arfit-package}
\title{arfit: Test for a trend in presence of autocorrelation}
\description{
Tools to test for a trend in presence of autocorrelation. All tools to
reproduce results in publication
}
\details{
\itemize{
\item Simulates data
\item Evaluates likelihood
\item Fits likelihood
\item Estimates MLE for model
\item Example provides
\item Tools for su=imulation study
}
}
\section{References}{


To learn more about using \code{arfit}, start with the vignette: \code{browseVignettes(package="arfit")} or click the index below
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/andybeet/arfit}
  \item Report bugs at \url{https://github.com/andybeet/arfit/issues}
}

}
\author{
\strong{Maintainer}: Andy Beet \email{andrew.beet@noaa.gov} (\href{https://orcid.org/0000-0001-8270-7090}{ORCID})

}
