% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_real_data.R
\name{fit_real_data}
\alias{fit_real_data}
\title{Run test on real data set}
\usage{
fit_real_data(dataSet, nBootSims = 499)
}
\arguments{
\item{dataSet}{Data frame. Two columns (x, y)}

\item{nBootSims}{Numeric scalar. Number of bootstrap samples to perform}
}
\description{
Supply data
}
