% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_ar2.R
\name{likelihood_ar2}
\alias{likelihood_ar2}
\title{Evaluates the likelihood of AR2}
\usage{
likelihood_ar2(data, beta, rho1, rho2)
}
\arguments{
\item{data}{dd}

\item{beta}{Numeric vector. intercept and slope term}

\item{rho1}{1st order Autoregressive parameter}

\item{rho2}{2nd order Autoregressive parameter}
}
\description{
Uses full likelihood as explained in Beach and MacKinnon
The first two data points are used in the likelihood
}
