% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_study_opt.R
\name{sim_study_opt}
\alias{sim_study_opt}
\title{Simulates n data sets and performs bootstrapping}
\usage{
sim_study_opt(
  outDir = here::here("out.txt"),
  betaVec = c(0, 0.12, 0.25, 0.5),
  rhoVec = c(0, 0.25, 0.5, 0.75, 0.95),
  sigmaVec = c(0.25, 0.5, 0.75),
  nTVec = c(10, 25, 50),
  nSims = 200,
  nBootSims = 500
)
}
\arguments{
\item{outDir}{Character string. Path to output file}

\item{betaVec}{Numeric vector. Values for beta_1 (slope/trend parameter)}

\item{rhoVec}{Numeric vector. Values for autoregressive parameter}

\item{sigmaVec}{Numeric vector. Values of standard deviation of noise}

\item{nSims}{Numeric scalar. Number of time series to simulate}

\item{nBootSims}{Numeric scalar. Number of bootstrap data sets}

\item{nVec}{Numeric vector. Values for the length of time series to simulate}
}
\description{
Performs n tests
}
\details{
For convenience the intercept, beta_0 is set to zero
}
\examples{
\dontrun{
}


}
